+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 ## 0.3.0
 
 * Added optional parameter azure_environment to provider config which defaults to public ([#437](https://github.com/databrickslabs/terraform-provider-databricks/pull/437)).
-* Added [databricks_service_principal](https://github.com/databrickslabs/terraform-provider-databricks/pull/386) resource.
+* * Added [databricks_service_principal](https://github.com/databrickslabs/terraform-provider-databricks/pull/386) resource.
+* `databricks_notebook` got behavior changed
 
 **Behavior changes**